Models are placed in separate files. Fixed errors caused in connection with this.
