+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,14 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Unwraper Man on 22.12.2022.
+//
+
+import Foundation
+
+struct QuizQuestion {
+    var image: String
+    var text: String
+    var correctAnswer: Bool
+}
