+++ MovieQuiz/Models/QuizResultViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizResultViewModel.swift
+//  MovieQuiz
+//
+//  Created by Unwraper Man on 22.12.2022.
+//
+
+import Foundation
+
+struct QuizResultViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
