+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Unwraper Man on 22.12.2022.
+//
+
+import UIKit
+import Foundation
+
+struct QuizStepViewModel {
+    var image: UIImage
+    var question: String
+    var questionNumber: String
+}
